Refactor to standardise on command pattern
