+++ test/extended/parse_code_test.rb
@@ -2,7 +2,7 @@ require "test_helper"
 
 module SnippetExtractor
   module Extended
-    class CodeParserTest < Minitest::Test
+    class ParseCodeTest < Minitest::Test
       def test_empty_syntax_trie_return_rules
         # Given
         syntax_trie = SyntaxTrie.new(Node.new({}, "", nil))
@@ -13,7 +13,7 @@ module SnippetExtractor
         expected = code
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_parsing_no_code_returns_no_code
@@ -35,7 +35,7 @@ module SnippetExtractor
         expected = ""
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_line_rule_whole_word_matches
@@ -64,7 +64,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_matches_ignore_case
@@ -93,7 +93,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_line_rule_whole_word_matches_at_end_of_line
@@ -123,7 +123,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_line_rule_partial_word_matches
@@ -148,7 +148,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_just_rule_whole_word_matches
@@ -178,7 +178,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_just_rule_partial_word_matches
@@ -204,7 +204,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_multi_rule_with_just_final
@@ -236,7 +236,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_multi_rule_with_line_final
@@ -269,7 +269,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_multi_rule_in_one_line_just_start_action
@@ -303,7 +303,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_multi_rule_in_one_line_line_start_action
@@ -336,7 +336,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_slight_overlapping_in_rules
@@ -363,7 +363,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_line_rule_skip_matches_inside_line
@@ -389,7 +389,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_multi_line_rule_skip_matches_inside
@@ -418,7 +418,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_repeat_char_wont_match_1_char
@@ -445,7 +445,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_repeat_char_default_is_tailless
@@ -475,7 +475,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie).join
       end
 
       def test_first_loc_argument
@@ -504,7 +504,7 @@ module SnippetExtractor
         CODE
 
         # Expect
-        assert_equal expected, CodeParser.new(code, syntax_trie, ['stop_at_first_loc']).parse.join
+        assert_equal expected, ParseCode.(code, syntax_trie, ['stop_at_first_loc']).join
       end
     end
   end
